Cannon 2.11.x (#1999)
* update cannon & router

* fix duplicated deps

* fix storage dumps

* fix error assertion reason usage

* update cannon to 2.11.7 and router to 3.3.7

* dedupe router

* fix spot-market tests

* upgrade hardhat-cannon

* Fix router version

* Remove unnecessary cannon cli dependency from root

* Add missing artifacts to cache

* More storage update

* update cannon to 2.11.8

* bump cannon (should be final)

* upgrade cannon to 2.11.10

---------

+++ auxiliary/OpGasPriceOracle/storage.dump.sol
@@ -19,7 +19,7 @@ contract PythStructs {
         int64 price;
         uint64 conf;
         int32 expo;
-        uint publishTime;
+        uint256 publishTime;
     }
     struct PriceFeed {
         bytes32 id;
